pyproject.toml
[project]
name = "chutes-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.13.2",
    "fastmcp>=2.13.0.2",
    "imagekitio>=4.2.0",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
    "loguru>=0.7.2",
]
